+++ repository/src/dao.rs
@@ -7,7 +7,7 @@ pub struct RosterItem {
     pub first_name: String,
     pub last_name: String,
     pub email: String,
-    pub salary: f64,
+    pub salary: BigDecimal,
 }
 
 #[derive(Debug, Clone, FromRow)]
